+++ src/main/java/com/raoulvdberge/refinedstorage/gui/GuiCraftingPreview.java
@@ -44,7 +44,7 @@ public class GuiCraftingPreview extends GuiBase {
     private List<ICraftingPreviewElement> stacks;
     private GuiScreen parent;
 
-    private ItemStack stack;
+    private int hash;
     private int quantity;
 
     private GuiButton startButton;
@@ -52,7 +52,7 @@ public class GuiCraftingPreview extends GuiBase {
 
     private IElementDrawers drawers = new CraftingPreviewElementDrawers();
 
-    public GuiCraftingPreview(GuiScreen parent, List<ICraftingPreviewElement> stacks, ItemStack stack, int quantity) {
+    public GuiCraftingPreview(GuiScreen parent, List<ICraftingPreviewElement> stacks, int hash, int quantity) {
         super(new Container() {
             @Override
             public boolean canInteractWith(EntityPlayer player) {
@@ -63,7 +63,7 @@ public class GuiCraftingPreview extends GuiBase {
         this.stacks = new ArrayList<>(stacks);
         this.parent = parent;
 
-        this.stack = stack;
+        this.hash = hash;
         this.quantity = quantity;
 
         this.scrollbar = new Scrollbar(149, 20, 12, 119);
@@ -184,7 +184,7 @@ public class GuiCraftingPreview extends GuiBase {
     }
 
     private void startRequest() {
-        RS.INSTANCE.network.sendToServer(new MessageGridCraftingStart(stack, quantity));
+        RS.INSTANCE.network.sendToServer(new MessageGridCraftingStart(hash, quantity));
 
         close();
     }