+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed device names overflowing Controller GUI (raoulvdberge)
 - Fixed high CPU load when Refined Storage GUIs are open (raoulvdberge)
 - Fixed not being able to extract Mekanism tanks and bins from the Grid (raoulvdberge)
+- Fixed not being able to craft Immersive Engineering Revolver (raoulvdberge)
 
 ### 1.5.18
 - Added Project E integration for the External Storage on the Transmutation Table (raoulvdberge)